q-reserve/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6

# Email and background jobs
celery==5.3.4
redis==5.0.1
aiosmtplib==3.0.1

# Templates and frontend
jinja2==3.1.2
aiofiles==23.2.1

# HTTP client for testing
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality and formatting
black==23.11.0
isort==5.12.0
ruff==0.1.6
pre-commit==3.6.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# File handling
python-magic==0.4.27
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2 